# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Langfuse Admin API
  description: |
    API de administración para Langfuse que permite gestionar organizaciones, proyectos, usuarios y API keys.
    Esta API está diseñada para uso administrativo con autenticación mediante API key.
  version: 1.0.0
  contact:
    name: Langfuse Team
    url: https://langfuse.com
servers:
  - url: http://localhost:3100/api
    description: Servidor local de desarrollo
  - url: https://admin-api.langfuse.com/api
    description: Servidor de producción
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Código de error
        message:
          type: string
          description: Mensaje de error
        details:
          type: object
          description: Detalles adicionales del error (opcional)
...
    ProjectMember:
      type: object
      properties:
        project_id:
          type: string
          description: ID del proyecto
        user_id:
          type: string
          description: ID del usuario
        role:
          type: string
          enum: [OWNER, ADMIN, MEMBER, VIEWER]
          description: Rol del usuario en el proyecto
        created_at:
          type: string
          format: date-time
          description: Fecha de creación
        updated_at:
          type: string
          format: date-time
          description: Fecha de última actualización
        name:
          type: string
          description: Nombre del usuario
        email:
          type: string
          format: email
          description: Email del usuario
        image:
          type: string
          description: URL de imagen del usuario
    User:
      type: object
      properties:
        id:
          type: string
          description: ID único del usuario
        name:
          type: string
          description: Nombre del usuario
        email:
          type: string
          format: email
          description: Email del usuario
        image:
          type: string
          description: URL de imagen del usuario
        created_at:
          type: string
          format: date-time
          description: Fecha de creación
        updated_at:
          type: string
          format: date-time
          description: Fecha de última actualización
        admin:
          type: boolean
          description: Si el usuario es administrador global
        feature_flags:
          type: array
          items:
            type: string
          description: Flags de funcionalidades habilitadas para el usuario
    UsersList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          type: object
          properties:
            total:
              type: integer
              description: Número total de usuarios
            page:
              type: integer
              description: Página actual
            limit:
              type: integer
              description: Límite de resultados por página
            pages:
              type: integer
              description: Número total de páginas
...
paths:
...
  /users:
    get:
      summary: Listar usuarios
      description: Obtiene todos los usuarios con paginación y búsqueda
      operationId: getUsers
      tags: [Users]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Término de búsqueda para nombre o email
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Límite de resultados por página
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Número de página
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersList'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Crear usuario
      description: Crea un nuevo usuario
      operationId: createUser
      tags: [Users]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  description: Nombre del usuario
                email:
                  type: string
                  format: email
                  description: Email del usuario
                password:
                  type: string
                  description: Contraseña (opcional)
                admin:
                  type: boolean
                  description: Si el usuario es administrador
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{id}:
    get:
      summary: Obtener usuario
      description: Obtiene un usuario por ID
      operationId: getUserById
      tags: [Users]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID del usuario
      responses:
        '200':
          description: Datos del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Actualizar usuario
      description: Actualiza un usuario existente
      operationId: updateUser
      tags: [Users]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nombre del usuario
                email:
                  type: string
                  format: email
                  description: Email del usuario
                image:
                  type: string
                  description: URL de imagen del usuario
                feature_flags:
                  type: array
                  items:
                    type: string
                  description: Flags de funcionalidades
                admin:
                  type: boolean
                  description: Si el usuario es administrador
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Eliminar usuario
      description: Elimina un usuario (verificando dependencias)
      operationId: deleteUser
      tags: [Users]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID del usuario
      responses:
        '200':
          description: Usuario eliminado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensaje de confirmación
        '400':
          description: No se puede eliminar (tiene dependencias)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /projects/{projectId}/members:
    get:
      summary: Listar miembros del proyecto
      description: Obtiene todos los miembros de un proyecto
      operationId: getProjectMembers
      tags: [Project Members]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
          description: ID del proyecto
      responses:
        '200':
          description: Lista de miembros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMember'
        '404':
          description: Proyecto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Añadir miembro al proyecto
      description: Añade un usuario a un proyecto
      operationId: addProjectMember
      tags: [Project Members]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
          description: ID del proyecto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
                  description: ID del usuario a añadir
                role:
                  type: string
                  enum: [OWNER, ADMIN, MEMBER, VIEWER]
                  default: VIEWER
                  description: Rol del usuario en el proyecto
      responses:
        '201':
          description: Usuario añadido al proyecto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Proyecto o usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: El usuario ya es miembro de este proyecto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{projectId}/members/batch:
    post:
      summary: Añadir múltiples miembros al proyecto
      description: Añade varios usuarios a un proyecto en una sola operación
      operationId: addBatchProjectMembers
      tags: [Project Members]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
          description: ID del proyecto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - members
              properties:
                members:
                  type: array
                  items:
                    type: object
                    required:
                      - userId
                    properties:
                      userId:
                        type: string
                        description: ID del usuario
                      role:
                        type: string
                        enum: [OWNER, ADMIN, MEMBER, VIEWER]
                        default: VIEWER
                        description: Rol del usuario
      responses:
        '201':
          description: Resultado de la operación por lotes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectMember'
                    description: Miembros añadidos correctamente
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        userId:
                          type: string
                        error:
                          type: string
                    description: Errores al añadir miembros
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Proyecto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{projectId}/members/{userId}:
    get:
      summary: Obtener miembro del proyecto
      description: Obtiene detalles de un miembro específico del proyecto
      operationId: getProjectMember
      tags: [Project Members]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
          description: ID del proyecto
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: ID del usuario
      responses:
        '200':
          description: Datos del miembro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
        '404':
          description: Membresía no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Actualizar miembro del proyecto
      description: Actualiza el rol de un miembro en el proyecto
      operationId: updateProjectMember
      tags: [Project Members]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
          description: ID del proyecto
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: ID del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum: [OWNER, ADMIN, MEMBER, VIEWER]
                  description: Nuevo rol del usuario
      responses:
        '200':
          description: Rol actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
        '400':
          description: Datos inválidos o intento de cambiar el rol del último propietario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Membresía no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Eliminar miembro del proyecto
      description: Elimina un miembro del proyecto
      operationId: removeProjectMember
      tags: [Project Members]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
          description: ID del proyecto
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: ID del usuario
      responses:
        '200':
          description: Miembro eliminado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensaje de confirmación
        '400':
          description: No se puede eliminar al último propietario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Membresía no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...